Change: Rename s/LotterySelect/LotteryView/g
