+++ src/component/LotteryView.js
